split out XHP definitions
