+++ guides/hack/21-XHP/10-interfaces-examples/all-in-one.inc.php
@@ -1,0 +1,7 @@
+<?hh // partial
+
+final class XHPUnsafeExample implements XHPUnsafeRenderable {
+  public function toHTMLString(): string {
+    return '<script>'.$_GET['I_LOVE_XSS'].'</script>';
+  }
+}
