+++ guides/hack/21-XHP/10-interfaces-examples/markdown-wrapper.inc.php
@@ -1,0 +1,14 @@
+<?hh // partial
+
+class ExampleMarkdownXHPWrapper implements XHPUnsafeRenderable {
+  private string $html;
+  public function __construct(string $markdown_source) {
+    $this->html = HHVM\UserDocumentation\XHP\Examples\md_render(
+      $markdown_source,
+    );
+  }
+
+  public function toHTMLString(): string {
+    return $this->html;
+  }
+}
