+++ guides/hack/21-XHP/10-interfaces-examples/markdown-wrapper.php
@@ -1,18 +1,5 @@
 <?hh // partial
 
-class ExampleMarkdownXHPWrapper implements XHPUnsafeRenderable {
-  private string $html;
-  public function __construct(string $markdown_source) {
-    $this->html = HHVM\UserDocumentation\XHP\Examples\md_render(
-      $markdown_source,
-    );
-  }
-
-  public function toHTMLString(): string {
-    return $this->html;
-  }
-}
-
 <<__EntryPoint>>
 function run(): void {
   \__init_autoload();
