+++ guides/hack/21-XHP/16-extending-examples/add-class.inc.php
@@ -1,0 +1,12 @@
+<?hh // partial
+
+class :my-cls-adder extends :x:element {
+  attribute :div;
+  use XHPHelpers;
+  protected function render(): \XHPRoot {
+    $div = <div />;
+    $div->addClass('my-cls-adder');
+    $div->appendChild($this->getChildren());
+    return $div;
+  }
+}
