+++ guides/hack/21-XHP/16-extending-examples/add-class.php
@@ -1,16 +1,5 @@
 <?hh // partial
 
-class :my-cls-adder extends :x:element {
-  attribute :div;
-  use XHPHelpers;
-  protected function render(): \XHPRoot {
-    $div = <div />;
-    $div->addClass('my-cls-adder');
-    $div->appendChild($this->getChildren());
-    return $div;
-  }
-}
-
 <<__EntryPoint>>
 function extending_examples_add_class_run(): void {
   \__init_autoload();
