+++ guides/hack/21-XHP/16-extending-examples/bad-attribute-transfer.inc.php
@@ -1,0 +1,11 @@
+<?hh // partial
+
+class :ui-my-box extends :x:element {
+  attribute :div; // inherit from attributes from <div>
+
+  protected function render(): \XHPRoot {
+    // returning this will ignore any attribute set on this custom object
+    // They are not transferred automatically when returning the <div>
+    return <div class="my-box">{$this->getChildren()}</div>;
+  }
+}
