+++ guides/hack/21-XHP/16-extending-examples/bad-attribute-transfer.php
@@ -1,15 +1,5 @@
 <?hh // partial
 
-class :ui-my-box extends :x:element {
-  attribute :div; // inherit from attributes from <div>
-
-  protected function render(): \XHPRoot {
-    // returning this will ignore any attribute set on this custom object
-    // They are not transferred automatically when returning the <div>
-    return <div class="my-box">{$this->getChildren()}</div>;
-  }
-}
-
 <<__EntryPoint>>
 function extending_examples_bad_attribute_transfer_run(): void {
   \__init_autoload();
