+++ guides/hack/21-XHP/16-extending-examples/categories.php
@@ -1,14 +1,5 @@
 <?hh // partial
 
-class :my-text extends :x:element {
-  category %phrase;
-  children (pcdata | %phrase); // prefixed colon ommitted purposely on pcdata
-
-  protected function render(): \XHPRoot {
-    return <x:frag>{$this->getChildren('%phrase')}</x:frag>;
-  }
-}
-
 <<__EntryPoint>>
 function extending_examples_categories_run(): void {
   \__init_autoload();
