+++ guides/hack/21-XHP/16-extending-examples/get-id.inc.php
@@ -1,0 +1,9 @@
+<?hh // partial
+
+class :my-id extends :x:element {
+  attribute string id;
+  use XHPHelpers;
+  protected function render(): \XHPRoot {
+    return <span id={$this->getID()}>This has a random id</span>;
+  }
+}
