+++ guides/hack/21-XHP/16-extending-examples/get-id.php
@@ -1,13 +1,5 @@
 <?hh // partial
 
-class :my-id extends :x:element {
-  attribute string id;
-  use XHPHelpers;
-  protected function render(): \XHPRoot {
-    return <span id={$this->getID()}>This has a random id</span>;
-  }
-}
-
 <<__EntryPoint>>
 function extending_examples_get_id_run(): void {
   \__init_autoload();
