+++ guides/hack/21-XHP/16-extending-examples/required-attributes.inc.php
@@ -1,0 +1,11 @@
+<?hh // partial
+
+class :user-info extends :x:element {
+  attribute int userid @required;
+  attribute string name = "";
+
+  protected function render(): \XHPRoot {
+    return
+      <x:frag>User with id {$this->:userid} has name {$this->:name}</x:frag>;
+  }
+}
