+++ guides/hack/21-XHP/16-extending-examples/required-attributes.php
@@ -1,15 +1,5 @@
 <?hh // partial
 
-class :user-info extends :x:element {
-  attribute int userid @required;
-  attribute string name = "";
-
-  protected function render(): \XHPRoot {
-    return
-      <x:frag>User with id {$this->:userid} has name {$this->:name}</x:frag>;
-  }
-}
-
 <<__EntryPoint>>
 function extending_examples_attributes_run(): void {
   \__init_autoload();
