+++ guides/hack/21-XHP/16-extending-examples/xhp-async.php
@@ -1,17 +1,5 @@
 <?hh // partial
 
-class :ui:get-status extends :x:element {
-
-  use XHPAsync;
-
-  protected async function asyncRender(): Awaitable<\XHPRoot> {
-    $ch = curl_init('https://developers.facebook.com/status/');
-    curl_setopt($ch, CURLOPT_USERAGENT, 'hhvm/user-documentation example');
-    $status = await HH\Asio\curl_exec($ch);
-    return <x:frag>Status is: {$status}</x:frag>;
-  }
-}
-
 <<__EntryPoint>>
 async function extending_examples_async_run(): Awaitable<void> {
   \__init_autoload();
